removed 'integrators.py'
